Create upgrade_common file and rename upgrade check
