+++ src/test/regress/upgrade/README.md
@@ -32,13 +32,13 @@ To see full command list:
 How the upgrade test works:
 - Temporary folder `tmp_upgrade` is created in `src/test/regress/`, if one exists it is removed first.
 - Database is initialized and citus cluster is created(1 coordinator + 2 workers) with old postgres.
-- `before_upgrade_schedule` is run with `pg_regress`. This schedule does not drop any tables or data so that we can verify upgrade.
+- `before_pg_upgrade_schedule` is run with `pg_regress`. This schedule does not drop any tables or data so that we can verify upgrade.
 - `citus_prepare_pg_upgrade` is run in coordinators and workers.
 - Old database is stopped.
 - A new database is initialized with new postgres under `tmp_upgrade`.
 - Postgres upgrade is performed.
 - New database is started in both coordinators and workers.
 - `citus_finish_pg_upgrade` is run in coordinators and workers to finalize the upgrade step.
-- `after_upgrade_schedule` is run with `pg_regress` to verify that the previously created tables, and data still exist. Router and realtime queries are used to verify this.
+- `after_pg_upgrade_schedule` is run with `pg_regress` to verify that the previously created tables, and data still exist. Router and realtime queries are used to verify this.
