+++ src/test/regress/upgrade/config.py
@@ -1,5 +1,5 @@
-BEFORE_UPGRADE_SCHEDULE = './before_upgrade_schedule'
-AFTER_UPGRADE_SCHEDULE = './after_upgrade_schedule'
+BEFORE_UPGRADE_SCHEDULE = './before_pg_upgrade_schedule'
+AFTER_UPGRADE_SCHEDULE = './after_pg_upgrade_schedule'
 
 BEFORE_CITUS_UPGRADE_SCHEDULE = './before_citus{}_upgrade_schedule'
 AFTER_CITUS_UPGRADE_SCHEDULE = './after_citus_upgrade_schedule'
@@ -17,7 +17,7 @@ class CitusUpgradeConfig():
         }
 
 
-class Config():
+class PGUpgradeConfig():
     def __init__(self, arguments):
         self.old_bindir = arguments['--old-bindir']
         self.new_bindir = arguments['--new-bindir']
